requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.1
click==8.2.1
conformer==0.3.2
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.5.1
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.2
joblib==1.5.1
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.0
omegaconf==2.3.0
onnx==1.18.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.2.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
resampy==0.4.3
resemble-perth==1.0.1
rich==14.0.0
ruff==0.11.12
s3tokenizer==0.1.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.23.2
virtualenv==20.31.2
watchfiles==1.0.4
websockets==15.0.1
zipp==3.22.0
